requirements.txt
streamlit
yt-dlp
python-dotenv
